+++ client/api/src/main/java/io/smallrye/graphql/client/core/Buildable.java
@@ -1,0 +1,7 @@
+package io.smallrye.graphql.client.core;
+
+import io.smallrye.graphql.client.core.exceptions.BuildException;
+
+public interface Buildable {
+    String build() throws BuildException;
+}
