+++ client/api/src/main/java/io/smallrye/graphql/client/core/VariableType.java
@@ -1,6 +1,6 @@
-package io.smallrye.graphql.client.mpapi.core;
+package io.smallrye.graphql.client.core;
 
-import static io.smallrye.graphql.client.mpapi.core.utils.ServiceUtils.getNewInstanceOf;
+import static io.smallrye.graphql.client.core.utils.ServiceUtils.getNewInstanceOf;
 
 public interface VariableType extends Buildable {
