+++ client/api/src/main/java/io/smallrye/graphql/client/core/exceptions/BuildException.java
@@ -1,4 +1,4 @@
-package io.smallrye.graphql.client.mpapi.core.exceptions;
+package io.smallrye.graphql.client.core.exceptions;
 
 public class BuildException extends RuntimeException {
     public BuildException(String msg) {
