--- client/api/src/main/java/io/smallrye/graphql/client/mpapi/core/Buildable.java
@@ -1,7 +1,0 @@
-package io.smallrye.graphql.client.mpapi.core;
-
-import io.smallrye.graphql.client.mpapi.core.exceptions.BuildException;
-
-public interface Buildable {
-    String build() throws BuildException;
-}
