+++ client/implementation-vertx/src/main/java/io/smallrye/graphql/client/dynamic/vertx/VertxDynamicGraphQLClient.java
@@ -12,16 +12,16 @@ import javax.json.JsonReader;
 import javax.json.bind.Jsonb;
 import javax.json.bind.JsonbBuilder;
 
+import io.smallrye.graphql.client.Error;
+import io.smallrye.graphql.client.Request;
+import io.smallrye.graphql.client.Response;
+import io.smallrye.graphql.client.core.Document;
 import io.smallrye.graphql.client.dynamic.ErrorImpl;
 import io.smallrye.graphql.client.dynamic.RequestImpl;
 import io.smallrye.graphql.client.dynamic.ResponseImpl;
 import io.smallrye.graphql.client.dynamic.SmallRyeGraphQLDynamicClientLogging;
 import io.smallrye.graphql.client.dynamic.SmallRyeGraphQLDynamicClientMessages;
 import io.smallrye.graphql.client.dynamic.api.DynamicGraphQLClient;
-import io.smallrye.graphql.client.mpapi.Error;
-import io.smallrye.graphql.client.mpapi.Request;
-import io.smallrye.graphql.client.mpapi.Response;
-import io.smallrye.graphql.client.mpapi.core.Document;
 import io.smallrye.mutiny.Uni;
 import io.vertx.core.MultiMap;
 import io.vertx.core.Vertx;
