+++ client/implementation/src/main/resources/META-INF/services/io.smallrye.graphql.client.core.Document
